Refactor: move error handler to api_utils.py
